Refactorización de tipos de datos: Migración de identificadores de Int a Long en entidades de inventario y horarios
* **Refactorización de Entidades:** Actualización de los campos `id`, `usuarioUid` y `articuloId` de `Int` a `Long` en `InventarioEntity` y `HorarioEntity` para asegurar la compatibilidad con identificadores de mayor precisión.
* **Ajustes en Mappers:** Eliminación de conversiones `.toInt()` en `InventarioMapper` y `HorarioMapper`, adaptando el flujo de datos desde los DTOs hacia las entidades actualizadas.
* **Lógica de Sincronización:** Actualización en `SessionSyncManager` para realizar el cast explícito a `Int` al consultar `PremiosRegistry` y mejoras en la legibilidad de condiciones lógicas mediante el uso de paréntesis.
* **Base de Datos y Debugging:** Adición de un log informativo en `Database.kt` para rastrear la cantidad de elementos procesados durante la sincronización del inventario local.
* **Consistencia en Red y Auth:** Refactorización menor de expresiones booleanas en `AuthRepositoryImpl` y `NetworkModule` para mejorar la claridad en la validación de tokens y estados de sesión.

+++ composeApp/src/commonMain/kotlin/pe/breaker/dkaviplay/data/entity/HorarioEntity.kt
@@ -1,8 +1,8 @@
 package pe.breaker.dkaviplay.data.entity
 
 data class HorarioEntity(
-    val id:Int,
-    val usuarioUid: Int,
+    val id: Long,
+    val usuarioUid: Long,
     val dia: Int,
     val horaInicio: String,
     val horaFin: String,
